+++ management/model_gateway_credential.go
@@ -12,6 +12,7 @@ package management
 
 import (
 	"encoding/json"
+	"time"
 )
 
 // GatewayCredential struct for GatewayCredential
@@ -19,7 +20,7 @@ type GatewayCredential struct {
 	// A string that specifies the auto-generated ID for this credential. This is the JWT's jti claim. This is a required property.
 	Id *string `json:"id,omitempty"`
 	// A date that specifies the date the credential was created in Coordinated Universal Time (UTC). This is a required property.
-	CreatedAt *string `json:"createdAt,omitempty"`
+	CreatedAt *time.Time `json:"createdAt,omitempty"`
 	// A date that specifies the date the credential was last used in UTC. This is a required property.
 	LastUsedAt *string `json:"lastUsedAt,omitempty"`
 	// A string that specifies the signed JWT for the gateway credential. This property is present only when the gateway credential is created.
@@ -76,9 +77,9 @@ func (o *GatewayCredential) SetId(v string) {
 }
 
 // GetCreatedAt returns the CreatedAt field value if set, zero value otherwise.
-func (o *GatewayCredential) GetCreatedAt() string {
+func (o *GatewayCredential) GetCreatedAt() time.Time {
 	if o == nil || o.CreatedAt == nil {
-		var ret string
+		var ret time.Time
 		return ret
 	}
 	return *o.CreatedAt
@@ -86,7 +87,7 @@ func (o *GatewayCredential) GetCreatedAt() string {
 
 // GetCreatedAtOk returns a tuple with the CreatedAt field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *GatewayCredential) GetCreatedAtOk() (*string, bool) {
+func (o *GatewayCredential) GetCreatedAtOk() (*time.Time, bool) {
 	if o == nil || o.CreatedAt == nil {
 		return nil, false
 	}
@@ -102,8 +103,8 @@ func (o *GatewayCredential) HasCreatedAt() bool {
 	return false
 }
 
-// SetCreatedAt gets a reference to the given string and assigns it to the CreatedAt field.
-func (o *GatewayCredential) SetCreatedAt(v string) {
+// SetCreatedAt gets a reference to the given time.Time and assigns it to the CreatedAt field.
+func (o *GatewayCredential) SetCreatedAt(v time.Time) {
 	o.CreatedAt = &v
 }